+++ Echo/EchoPlugin.cs
@@ -64,6 +64,8 @@ namespace Echo
             AlphaPlugin.CommandManager?.Register(new EchoToggleCommand());
             AlphaPlugin.CommandManager?.Register(new EchoTeleportPersonCommand());
             AlphaPlugin.CommandManager?.Register(new EchoTeleportLocationCommand());
+            AlphaPlugin.CommandManager?.Register(new EchoLocationListCommand());
+            AlphaPlugin.CommandManager?.Register(new EchoLocationRemoveCommand());
             AlphaPlugin.CommandManager?.Register(new EchoCopyNameCommand());
             AlphaPlugin.CommandManager?.Register(new EchoRevertNameCommand());
             AlphaPlugin.CommandManager?.Register(new EchoCopyOutfitCommand());